feat: Extend database schema and service functionality for customer and inventory management
- Add quantity column to inventory_items table for better stock tracking
- Implement get_random_customer and customer_exists methods in CustomerRepository for enhanced customer retrieval
- Introduce get_random_item and item_exists methods in InventoryRepository to support random item selection
- Update CustomerService and InventoryService to include new methods for customer and item existence checks
- Refactor OrderProducerService to utilize customer and inventory services for order generation, ensuring valid customer and item selection

+++ logistics-engine/src/api/handlers/customer_handlers.rs
@@ -20,7 +20,11 @@ pub async fn list_customers(
     println!("pagination: {:?}", pagination);
     let customers = state
         .customer_service
-        .get_all_customers(pagination.page, pagination.limit, pagination.search.clone())
+        .get_all_customers(
+            pagination.page as usize,
+            pagination.limit as usize,
+            pagination.search.clone(),
+        )
         .await?;
 
     Ok(success(customers))
